robotic-hand-autonomous-grasping/requirements.txt
opencv-python
pyserial
torch
torchvision
scikit-learn
numpy 